Team Refactor / Clean Up
