+++ src/routes/profile/index.ts
@@ -1,10 +1,8 @@
 import { createRoute, OpenAPIHono } from '@hono/zod-openapi';
-import type { Context, Next } from 'hono';
 import { z } from 'zod';
 import { DBTables } from '../../constants/db.ts';
+import { canModifyProfile } from '../../middleware/canModifyProfile.ts';
 import { authorizeAdmin, authorizeOrganizer, authorizeVolunteer } from '../../middleware/createMiddleware.ts';
-import { Access } from '../../types/data/access.ts';
-import type { SessionData } from '../../types/data/session';
 import {
 	type DataResponse,
 	generateDataResponeSchema,
@@ -150,23 +148,6 @@ protectedProfileRoutes.openapi(
 		return context.json({ message: 'Profile created successfully' } satisfies StatusResponse, StatusCodes.CREATED);
 	}
 );
-
-async function canModifyProfile(context: Context, next: Next) {
-	const session = context.get('session') as SessionData;
-	const targetId = context.req.param('id');
-
-	// If admin, allow
-	if (session.access === Access.ADMIN) {
-		return next();
-	}
-
-	// For volunteers, only allow if it's their own profile
-	if (session.id !== targetId) {
-		return context.json({ message: 'Can only modify own profile' }, StatusCodes.FORBIDDEN);
-	}
-
-	return next();
-}
 
 // PATCH profile
 protectedProfileRoutes.openapi(
